# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RL-A2A v2.0
  description: Reinforcement Learning Agent-to-Agent Communication Platform
  version: 2.0.0
paths:
...
  /api/v1/messages/send:
    post:
      tags:
      - messages
      summary: Send Message
      description: Send a message (REST endpoint)
      operationId: send_message_api_v1_messages_send_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageSendRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/messages/jsonrpc:
    post:
      tags:
      - messages
      summary: Jsonrpc Endpoint
      description: JSON-RPC 2.0 endpoint
      operationId: jsonrpc_endpoint_api_v1_messages_jsonrpc_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONRPCRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JSONRPCRequest:
      properties:
        jsonrpc:
          type: string
          title: Jsonrpc
          default: '2.0'
        method:
          type: string
          title: Method
        params:
          additionalProperties: true
          type: object
          title: Params
        id:
          anyOf:
          - {}
          - type: 'null'
          title: Id
      type: object
      required:
      - method
      - params
      title: JSONRPCRequest
...
    MessageSendRequest:
      properties:
        sender_id:
          type: string
          title: Sender Id
        receiver_id:
          type: string
          title: Receiver Id
        content:
          title: Content
        message_type:
          type: string
          title: Message Type
          default: text
        priority:
          type: integer
          title: Priority
          default: 2
      type: object
      required:
      - sender_id
      - receiver_id
      - content
      title: MessageSendRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...